71 Refactoring edit group of lessons and add min & max in DateTimeRange class
