+++ src/main/java/com/school/school/lessons_groups/LessonsGroupRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LessonsGroupsRepository extends JpaRepository<LessonsGroups, Long>, CustomRepository<LessonsGroups> {
+public interface LessonsGroupRepository extends JpaRepository<LessonsGroup, Long>, CustomRepository<LessonsGroup> {
     boolean existsByTitle(String title);
     boolean existsByTitleAndIdNot(String title, long id);
 }
